requirements.txt
setuptools~=80.9.0
beautifulsoup4~=4.13.4
pyyaml~=6.0.2
bertopic~=0.17.3
scikit-learn~=1.7.1
tqdm~=4.67.1
requests~=2.32.4